+++ betfairlightweight/streaming/listener.py
@@ -16,6 +16,7 @@ class BaseListener:
         self.stream = None
         self.stream_type = None  # marketSubscription/orderSubscription
         self.stream_unique_id = None
+        self.connections_available = None  # connection throttling
 
     def register_stream(self, unique_id: int, operation: str) -> None:
         logger.info("Register: %s %s" % (operation, unique_id))
@@ -137,14 +138,19 @@ class StreamListener(BaseListener):
             "[Connect: %s]: connection_id: %s" % (unique_id, self.connection_id)
         )
 
-    @staticmethod
-    def _on_status(data: dict, unique_id: int) -> None:
+    def _on_status(self, data: dict, unique_id: int) -> None:
         """Called on status operation
 
         :param data: Received data
         """
         status_code = data.get("statusCode")
-        logger.info("[Subscription: %s]: %s" % (unique_id, status_code))
+        connections_available = data.get("connectionsAvailable")
+        if connections_available:
+            self.connections_available = data.get("connectionsAvailable")
+        logger.info(
+            "[Subscription: %s]: %s (%s connections available)"
+            % (unique_id, status_code, self.connections_available)
+        )
 
     def _on_change_message(self, data: dict, unique_id: int) -> None:
         change_type = data.get("ct", "UPDATE")