+++ HISTORY.rst
@@ -3,6 +3,19 @@
 Release History
 ---------------
 
+2.1.0 (2020-03-02)
++++++++++++++++++++
+
+**Improvements**
+
+- datetime handling added to time_range filter (@trigvi)
+- connectionsAvailable handling added
+
+**Bug Fixes**
+
+- #273 error handling added for markets without marketDefinition
+- #233 sendall used instead of send so that all data is sent (bug present since 2016!)
+
 2.0.1 (2020-02-17)
 +++++++++++++++++++